+++ src/ArduinoJson/Strings/JsonString.hpp
@@ -4,6 +4,7 @@
 
 #pragma once
 
+#include <ArduinoJson/Polyfills/type_traits.hpp>
 #include <ArduinoJson/Strings/Adapters/RamString.hpp>
 
 #if ARDUINOJSON_ENABLE_STD_STREAM
@@ -18,15 +19,16 @@ class JsonString {
   friend struct detail::StringAdapter<JsonString>;
 
  public:
-  enum Ownership { Copied, Linked };
-
   JsonString() : str_(nullptr, 0, true) {}
 
-  JsonString(const char* data, Ownership ownership = Copied)
-      : str_(data, data ? ::strlen(data) : 0, ownership == Linked) {}
+  JsonString(const char* data, bool isStatic = false)
+      : str_(data, data ? ::strlen(data) : 0, isStatic) {}
 
-  JsonString(const char* data, size_t sz, Ownership ownership = Copied)
-      : str_(data, sz, ownership == Linked) {}
+  template <typename TSize, typename = detail::enable_if_t<
+                                detail::is_integral<TSize>::value &&
+                                !detail::is_same<TSize, bool>::value>>
+  JsonString(const char* data, TSize sz, bool isStatic = false)
+      : str_(data, size_t(sz), isStatic) {}
 
   // Returns a pointer to the characters.
   const char* c_str() const {